Fixed how an edited message behaves
Allows for bot to remove it's second message,, and correctly behaves for moving free text around.
***does not allow a second message to be re-added if you remove it***
intentional- so you can't have the bot randomly sending messages *much* later.

Played with idea of editing an old message and adding a link into it; but it seemed to cause a raid bot due to strange behavior of edit (or my missunderstanding of how edit works)